cleanup(libscap): move scap_gethostname
